pyproject.toml
[project]
name = "yolo-obs"
version = "0.1.0"
description = "A simple Python program that allows the integration of YOLO object detection with OBS"
authors = [
    {name = "ChiefBacon",email = "tandcclarkkids@gmail.com"}
]

license = {text = "MPL-2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "obs-websocket-py (>=1.0,<2.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "ultralytics (>=8.3.127,<9.0.0)",
    "requests (>=2.28.1,<3.0.0)",
    "rich>=14.1.0",
]

[project.optional-dependencies]
cpu = [
    "torch>=2.7.0",
    "torchvision>=0.22.0"
]
gpu = [
    "torch>=2.7.0",
    "torchvision>=0.22.0"
]

[tool.uv]

# Prevent user from installing both extras together
conflicts = [
    [
        { extra = "cpu" },
        { extra = "gpu" }
    ]
]

[tool.uv.sources]
torch = [
    { index = "torch-gpu", extra = "gpu" }
]
torchvision = [
    { index = "torch-gpu", extra = "gpu" }
]

[[tool.uv.index]]
name = "torch-gpu"
url = "https://download.pytorch.org/whl/cu128"
explicit = true